[RHTAP-1078] Removing code duplication, renaming, moving files
