+++ tests/release/release.go
@@ -53,11 +53,11 @@ var _ = framework.ReleaseSuiteDescribe("[HACBS-1108]test-release-service-happy-p
 		Expect(err).NotTo(HaveOccurred(), "Error when fetching the default ECP: %v", err)
 		ecPolicy = defaultECP.Spec
 
-		_, err = fw.AsKubeAdmin.IntegrationController.CreateSnapshotWithComponents(applicationName, devNamespace, snapshotName, snapshotComponents)
+		_, err = fw.AsKubeAdmin.IntegrationController.CreateSnapshotWithComponents(snapshotName, "", applicationName, devNamespace, snapshotComponents)
 		Expect(err).NotTo(HaveOccurred())
 		_, err = fw.AsKubeAdmin.ReleaseController.CreateReleaseStrategy(releaseStrategyName, managedNamespace, releasePipelineName, releasePipelineBundle, releaseStrategyPolicy, serviceAccount, paramsReleaseStrategyM6)
 		Expect(err).NotTo(HaveOccurred())
-		_, err = fw.AsKubeAdmin.ReleaseController.CreateReleasePlanWithLabel(sourceReleasePlanName, devNamespace, applicationName, managedNamespace, "")
+		_, err = fw.AsKubeAdmin.ReleaseController.CreateReleasePlan(sourceReleasePlanName, devNamespace, applicationName, managedNamespace, "")
 		Expect(err).NotTo(HaveOccurred())
 		_, err = fw.AsKubeAdmin.TektonController.CreateEnterpriseContractPolicy(releaseStrategyPolicy, managedNamespace, ecPolicy)
 		Expect(err).NotTo(HaveOccurred())
